Updates needed to simulating surv times based on multilevel longitudinal data
Needed to allow separate pars (parameters) data frames to be passed to
the hazfn. This is because it is not always possible to pass all
parameters relating to one individual in a single row of the data frame.
For example, when there is multilevel clustered data, there may need to
be more than one row of parameters for each individual.